limited_memory_location
union is similar to structure except it has LIMITED memory . we will access or initialize the variable of union using its datatype
variable just like struture. it doesnt matter how many variables are declared in union , we can store the value of only one variable in it.ex-ankush is the datatype variable of union and rollno &age are the variables of union.among ankush.age and ankush.rollno only one value will be printed . with one datatype variable you can INITIALIZE only one local variable of union because the memory is LIMITED and  simply because of the property of union. one datatype variable-can INITIALIZE- only one local variable in union.
union are outdated , not much in use.