freeing_the_pointer
Dangling pointer is the pointer whose value has been deleted at the location but still points to the same memory location. by using delete keyword you can just delete the value, the pointer points to in this case is deleting the value 10(any value) but it's still pointing to same location which is called dangling pointer , in order to remove this pointing, we use null pointer or assign null value to the pointer so it removes the pointing to that location in memory.