Dynamic_allocation_of_array_values
we can allocate array values dynamically . we will asign the address of the array through "new" and store it in the pointer.look for the program. The values can be assigned using for loop .ex- *(pointer +i); is referring to the (0+i)th index of the array .look for the program.