Function_calling_itself
The function that calls itself is the recursive function. The common  example is finding factorial. logic is important . Consider it as different windows(functions) open at the same time. The result of last window is used for 2nd last window . Similarly , the result of 2nd last window will be used as the result of 3rd last window. It will go on  until the rsult of '2nd window' is used for'1st window' or the 'function' . The function is calling  itself until the termination condition is met.