Introduction_to_classes
Classes are similar to the structures. They are user difined datatype containing variables , methods and is called in the main fn using the variable of the class. But there are some differences. It can have children classes that can acquire the properties of parent classes . A class has all members private by default. A struct is a class where members are public by default. ex-class student{ };  int main( ){ student anil; }