different_ways_to_create_objects_of_classes(specially pointer_object)
we can use "pointer objects" of classes also .The values will be stored in the heap rather than stack. syntax- Human *bunty= new Human();//here , *bunty is the pointer of human. it is used to refer many memory locations of the class human.
bunty->name="bunty";//so, object bunty of "human" class is referring to the memory location containing the name "bunty"