PrivateAccessSpecifier_in_cpp
Private Access Specifier is used when we dont want to share the data or the working of data to all the users. Private data can be accessed through Public specifier by acquiring the properties of private in the public. Then , when we will call the public specifier properties(methods,fns or variables) in our main function through the class object, we will indirectly acquire the properties of the private accessifier.