Constructors
contructor---- It has same name as of class and it does not have any return type. It will be called AUTOMATICALLY when class's object is created in the main function and all its statements and methods will be executed automatically. it is mainly used to store default values , methods that we want to always execute  when class's object is called.it is generally declared in public specifier.
syntax-class Human{public:
						Human(){methods}};//name of constructor should exactly be same