Calling_constructor_default_values
If the constructor takes the parameter, than we can declare the value of parameter while declaring th object of class or during defining the constrcutor in class. syntax-Human rekha("roni");//we could write it without parameters also if values are declared during initialization in the constructor.in this example name of person was defined during declaring the object of class but age wasnt . then age would be the value declared during defining the constructor.take a look at code.