~destructor
Destructor is used to destroy the constructor.  it releases the memory taken by constructor members(variables,functions etc).we can use destructor only if the class object is a pointer.then we will delete the object of the class using "delete" operator which will call  the destructor.REMEMBER, we are only calling the destuctor through "delete" operator. the destructor should contain the other delete items to free the memory.syntax- ~Human(){ };