updation_in_values
Static variable or member is the one whose value changes after each update if updation is done. Each time you access the static variable you will get the updated value of static variable .if the static variable is in the class,then we have to define it outside the class using SCOPE operator. look at the code.