Friend_fn_and_class
Friend function is the special type of fn that can access the member of the class whose friend it is. It will be declared inside the class but it will be defined outside the class. It can also access  private as well as protected members of the class along with public members.