Acquiring_properties
Acquiring the properties of one class into another is called inheritance. the class which is inherited or gets inherited is called base class and the class which inherits the properties is called derieved or subclass. the subclass can access the members of the base class without "dot" operator. ex-Human.name="asdf"; is not neccessary rather it can directly access the member.ex-name=asdf"; .Look for code for better understanding .