Order_of_Execution_of_Constructors_and_Destructors_in_Inheritance
While calling the constructor and destructor of the base class through the object of derived class which also has its constructor and destructor, it will first call the C of the base class, then C of the derived class and then it will call the D of DERIVED class and later, it will call D of base class. IF pointer object is used for the class , then we will use delete keyword to call the destructor.