ability_of_an_object_to_take_on_many_forms
it means many forms. it occurs when there is heirarchy of classes with same functions (same name and parameters, identical) and we have to decide which function of which class has to be called and which function has to be OVERRIDDEN.the fn can be overidden using VIRTUAL keyword. it will used infront the fn. syntax :- virtual void introduce(){ }; 
-in the program ,we used "passed by  reference" for the local variable , its explanation is in other file.
for rememberance, remember that polymorphism involves overriding of a function.