Virtual Functions_Inherited_Attributes,_Hierarchical_Nature
If a function of certain class(generally base class) is declared as virtual , then as per rules, it will be OVERRIDDEN if the ditto or identical function is available in the derived class. BUT if the fn is not available in the derived class , then the base class fn is called. so for definition:- virtual function decides which fn is going to be OVERRIDDEN to allow the other identical fn of different class to get executed.