Pure_Virtual_Functions_Abstract_Classes
A function is said to be pure if it HAS been defined in the base class. It has to be defined in the base class. there is no backup of calling the (identical) fn from the base class as we did earlier.if we dont DEFINE this method in the base class, then it will show error. so the function being PURE just insures that the fn has been definitely defined in the derived class.
the pure function can also be defined but its not of much use since the fn is made PURE to make sure that the identical fn is defined in the derived class. We also cant call it from base class's own object because it has become an abstract class and abstract classes cannot have their objects.
-an ABSTRACT class is the one which has one or more PURE functions declared in it . an abstract class cannot have its own object.