Nested_Classes_or_Inner_classes
Nested class:- A class can be defined inside another class. it is known as nested class. nested class object can only be declared in the parent class , and not in any other function like main fn.we can declare the nested class object in the main fn using the object of class fn and scope resolution and the syntax is ---syntax:- Person::Address ad; where "Person" is the parent or main class and "Address" is the nested class.and "ad" is the object of Address class.