function_containing_the_"throw"_which_passed_in_"try"_block
In "fuctions throwing exception", "try" block has a "function" instead of "throw" keyword. The function will contain the "throw" keyword and the object will be passed to the "catch" block.  syntax :-  void test() throw(int,char,runtime_error)//datatype of object is mentioned in the scope. it is neccessary to write it.
{
throw 'c';

};